fix: resolve 5 critical + 2 medium bugs from code audit
Critical fixes:
- execution.rs: subscribe to Topic::Intent to capture PortfolioIntent,
  look up original intent by decision.intent_id instead of using
  UNKNOWN placeholders and zero quantity
- risk.rs: subscribe to Topic::Execution, decrement exposure on
  sell fills and position closes (was monotonically increasing)
- portfolio.rs: track capital_reserved per intent, return capital
  on RiskDecision rejection and ExecutionReport failure
- cycle_manager.rs: use max_focus_set from config instead of
  hardcoded 10, merge must_monitor into focus set
- strategy.rs: clear active_signals on new cycle (Scan stage),
  remove instruments on execution completion

Medium fixes:
- risk.rs: use f64 comparison for concentration/drawdown checks
  instead of truncating via as u8 cast
- execution.rs: set avg_fill_price from venue's last_fill_price()
  on filled orders, add last_fill_price to VenueAdapter trait

+++ crates/bamboo-core/src/config.rs
@@ -69,8 +69,8 @@ pub struct RiskLimitsConfig {
     pub max_position_size_usd: Money,
     #[serde(deserialize_with = "deserialize_usd_money")]
     pub max_portfolio_exposure_usd: Money,
-    pub max_concentration_pct: u8,
-    pub max_drawdown_pct: u8,
+    pub max_concentration_pct: f64,
+    pub max_drawdown_pct: f64,
     pub order_rate_limit_per_min: u32,
     pub kill_switch_enabled: bool,
 }
